backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
marshmallow==3.20.1
bcrypt==4.1.1
python-dotenv==1.0.0
email-validator==2.1.0
